pre-cbt. build test

+++ Unity-Garden_Hero/Assets/Scripts/Battle/StunUIManager.cs
@@ -39,7 +39,7 @@ public class StunUIManager : MonoBehaviour
         monsterManager.OnStunBroken.AddListener(OnStunBroken);
     }
     
-    void OnStunChanged(int newStun)
+    void OnStunChanged(float newStun)
     {
         UpdateStunBar();
